backend/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database - Firebase
firebase-admin==6.5.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic-settings==2.1.0

# Dependency Injection
dependency-injector==4.41.0

# Redis (for caching and task queue)
redis==5.0.1
hiredis==2.2.3

# Background tasks
celery==5.3.4

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
faker==20.1.0

# Code quality
black==23.12.0
isort==5.13.2
ruff==0.1.8
mypy==1.7.1

# Monitoring
sentry-sdk[fastapi]==1.39.1

# Utilities
python-slugify==8.0.1
email-validator==2.1.0

# AI/Agent Services
openai==1.6.1
anthropic==0.8.1
chromadb==0.4.22
tiktoken==0.5.2
